refactor to service based installs and add cli support
