+++ WPILibInstaller-Avalonia/Interfaces/IShortcutService.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace WPILibInstaller.Interfaces
+{
+    public interface IShortcutService
+    {
+        Task RunShortcutCreator(CancellationToken token);
+    }
+}
